+++ lib/ref_inspector/downloader/readme.ex
@@ -1,22 +1,36 @@
 defmodule RefInspector.Downloader.README do
   @moduledoc false
 
+  require Logger
+
   alias RefInspector.Config
   alias RefInspector.Database.Location
 
   @readme "ref_inspector.readme.md"
 
-  @doc """
-  Returns the path to the local copy of the README file.
-  """
+  @doc false
   @spec path_local() :: binary
-  def path_local, do: Location.local(@readme)
+  def path_local do
+    _ =
+      Logger.info(
+        "RefInspector.Downloader.README.path_local/0 has been" <>
+          " declared internal and will eventually be removed."
+      )
 
-  @doc """
-  Returns the path of the README file distributed in priv_dir.
-  """
+    Location.local(@readme)
+  end
+
+  @doc false
   @spec path_priv() :: binary
-  def path_priv, do: Application.app_dir(:ref_inspector, ["priv", @readme])
+  def path_priv do
+    _ =
+      Logger.info(
+        "RefInspector.Downloader.README.path_priv/0 has been" <>
+          " declared internal and will eventually be removed."
+      )
+
+    Application.app_dir(:ref_inspector, ["priv", @readme])
+  end
 
   @doc """
   Writes the informational README file if remote database is the default.
@@ -33,14 +47,16 @@ defmodule RefInspector.Downloader.README do
   end
 
   defp do_write do
-    readme_local = path_local()
+    readme_local = Location.local(@readme)
+    readme_priv = Application.app_dir(:ref_inspector, ["priv", @readme])
+
     dirname_local = Path.dirname(readme_local)
 
     unless File.dir?(dirname_local) do
       File.mkdir_p!(dirname_local)
     end
 
-    {:ok, _} = File.copy(path_priv(), readme_local)
+    {:ok, _} = File.copy(readme_priv, readme_local)
     :ok
   end
 end