+++ test/ref_inspector/downloader/readme_test.exs
@@ -3,13 +3,15 @@ defmodule RefInspector.Downloader.READMETest do
 
   alias RefInspector.Downloader.README
 
+  @test_path Path.expand("../../downloads", __DIR__)
+  @test_readme Path.join(@test_path, "ref_inspector.readme.md")
+
   setup_all do
     database_path = Application.get_env(:ref_inspector, :database_path)
     remote_urls = Application.get_env(:ref_inspector, :remote_urls)
-    test_path = Path.expand("../../downloads", __DIR__)
 
-    :ok = Application.put_env(:ref_inspector, :database_path, test_path)
-    _ = File.rm(README.path_local())
+    :ok = Application.put_env(:ref_inspector, :database_path, @test_path)
+    _ = File.rm(@test_readme)
 
     on_exit(fn ->
       :ok = Application.put_env(:ref_inspector, :database_path, database_path)
@@ -21,17 +23,17 @@ defmodule RefInspector.Downloader.READMETest do
     :ok = Application.put_env(:ref_inspector, :remote_urls, ["non-default-remote.yml"])
     :ok = README.write()
 
-    refute File.exists?(README.path_local())
+    refute File.exists?(@test_readme)
 
     :ok = Application.delete_env(:ref_inspector, :remote_urls)
     :ok = Application.put_env(:ref_inspector, :skip_download_readme, true)
     :ok = README.write()
 
-    refute File.exists?(README.path_local())
+    refute File.exists?(@test_readme)
 
     :ok = Application.delete_env(:ref_inspector, :skip_download_readme)
     :ok = README.write()
 
-    assert File.exists?(README.path_local())
+    assert File.exists?(@test_readme)
   end
 end